requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
django-environ==0.8.1
djangorestframework==3.13.1
mysqlclient==2.1.0
pytz==2021.3
slack-sdk==3.15.2
sqlparse==0.4.2
